feat: create new file for home view
making index view (home) and the template too so u don't need to type the html code again just use the extend and section just like i do in index. and for the style u can use the layout style file if u using that link to all of ur project. if no u can add it manual on the project section named link.